#57: Refactored logging
