+++ .lintstagedrc.json
@@ -1,10 +1,4 @@
 {
-  "*.js": [
-    "npx prettier --write",
-    "npx eslint . --fix"
-  ],
-  "*.mjs": [
-    "npx prettier --write",
-    "npx eslint . --fix"
-  ]
+  "*.js": ["npx prettier --write", "npx eslint . --fix"],
+  "*.mjs": ["npx prettier --write", "npx eslint . --fix"]
 }
