+++ index.js
@@ -61,12 +61,12 @@ if (require?.main === module) {
     .option('-i, --clientSecret <string>', 'OAuth2 client_secret parameter, use this OR bearerToken')
     .option('-k, --tokenUri <string>', 'OAuth2 token_uri parameter, use this OR bearerToken')
     .option('-e, --scope <string>', 'Optional OAuth2 scopes for client credentials')
-    .option('-m, --pathToMetadataReportJson <string>', 'Path to metadata report JSON')
+    .option('-p, --pathToMetadataReportJson <string>', 'Path to metadata report JSON')
     .option('-r, --resourceName <string>', 'Resource name to replicate data from')
     .option('-x, --expansions <items>', 'Comma-separated list of items to expand during the query process, e.g. Media,OpenHouse')
     .option('-f, --filter <string>', 'OData $filter expression')
     .option('-t, --top <number>', 'Optional parameter to use for OData $top')
-    .option('-p, --maxPageSize <number>', 'Optional parameter for the odata.maxpagesize header')
+    .option('-m, --maxPageSize <number>', 'Optional parameter for the odata.maxpagesize header')
     .option('-o, --outputPath <string>', 'Name of directory for results')
     .option('-l, --limit <number>', 'Limit total number of records at client level')
     .option('-v, --version <string>', 'Data Dictionary version to use', '2.0')
