+++ lib/find-variations/index.js
@@ -1,10 +1,9 @@
 'use strict';
 
-const chalk = require('chalk');
 const { readFile, writeFile } = require('fs/promises');
 const { distance } = require('fastest-levenshtein');
 const { getReferenceMetadata } = require('reso-certification-etl');
-const { buildMetadataMap } = require('../../common');
+const { buildMetadataMap, getLoggers } = require('../../common');
 const { DEFAULT_DD_VERSION } = require('../replication/utils');
 const humanizeDuration = require('humanize-duration');
 
@@ -27,26 +26,6 @@ const MATCHED_ON = Object.freeze({
   LOOKUP_VALUE: 'lookupValue',
   LEGACY_ODATA_VALUE: 'legacyODataValue'
 });
-
-/**
- * Creates loggers
- * @param {Boolean} fromCli true if using the console, false otherwise (default)
- * @returns a pair of loggers, one for normal messages and another for errors
- */
-const getLoggers = (fromCli = false) => {
-  const noop = () => {};
-  if (fromCli) {
-    return {
-      LOG: message => console.log(message),
-      LOG_ERROR: message => console.error(chalk.redBright.bold(message))
-    };
-  } else {
-    return {
-      LOG: noop,
-      LOG_ERROR: noop
-    };
-  }
-};
 
 /**
  * Prepares a metadata report so it can be searched easily
@@ -515,15 +494,13 @@ const computeVariations = async ({
 
                         if (!hasStandardLegacyODataValue) {
                           if (isExactMatch || isSubstringMatch) {
-                            const { lookupValue: standardLookupValue, ddWikiUrl } =
+                            const { ddWikiUrl } =
                               standardMetadataMap?.[resourceName]?.[fieldName]?.legacyODataValues?.[standardODataLookupValue] || {};
 
                             const suggestion = {
                               resourceName,
                               fieldName,
-                              lookupValue,
                               legacyODataValue,
-                              suggestedLookupValue: standardLookupValue,
                               suggestedLegacyODataValue: standardODataLookupValue,
                               matchedOn: MATCHED_ON.LEGACY_ODATA_VALUE,
                               strategy: MATCHING_STRATEGIES.SUBSTRING,
@@ -587,7 +564,6 @@ const computeVariations = async ({
     });
 
     if (fromCli) {
-      console.log('Done!');
       console.log(`Time Taken: ${calculateElapsedTimeString(startTime)}`);
     }
 
@@ -624,21 +600,18 @@ const findVariations = async ({
     return;
   }
 
-  if (!parseFloat(fuzziness) || fuzziness < 0 || fuzziness > 1) {
+  if (fuzziness < 0 || fuzziness > 1) {
     LOG_ERROR('Invalid value! fuzziness must be a decimal number in the range [0, 1]');
     return;
   }
 
-  LOG(`Using fuzziness of up to ${Math.round(fuzziness * 100)}% of word length!`);
+  LOG(`\nUsing fuzziness of up to ${Math.round(fuzziness * 100)}% of word length!\n`);
 
-  const TOTAL_START_TIME = new Date();
+  const START_TIME = new Date();
 
   try {
-    //load metadata report from given path - might nee to take report json rather than from path
-    LOG(`Loading metadata report: ${pathToMetadataReportJson}`);
-
     const metadataReportJson = JSON.parse(await readFile(pathToMetadataReportJson, { encoding: 'utf8' }));
-    LOG('Done!');
+    LOG(`Using metadata report: ${pathToMetadataReportJson}`);
 
     const args = {
       metadataReportJson,
@@ -654,9 +627,8 @@ const findVariations = async ({
 
     const report = await computeVariations(args);
 
-    LOG(`Saving results to ${VARIATIONS_RESULTS_FILE}...`);
     await writeFile(VARIATIONS_RESULTS_FILE, Buffer.from(JSON.stringify(report, null, 2)));
-    LOG('Done!');
+    LOG(`Results saved to '${VARIATIONS_RESULTS_FILE}'`);
 
     LOG(
       [
@@ -666,8 +638,8 @@ const findVariations = async ({
         `  • Suggested Lookups: ${report?.variations?.lookups.length ?? 0}`,
         `  • Suggested Expansions: ${report?.variations?.expansions?.length ?? 0}`,
         `  • Suggested Complex Types: ${report?.variations?.complexTypes?.length ?? 0}`,
-        'Processing complete! Exiting...',
-        `Total runtime: ${calculateElapsedTimeString(TOTAL_START_TIME)}`
+        '\nProcessing complete!\n',
+        `Total runtime: ${calculateElapsedTimeString(START_TIME)}\n`
       ].join('\n')
     );
   } catch (err) {
