+++ lib/misc/data-access/cert-api-client.js
@@ -3,19 +3,13 @@
 // TODO: change this to use the Fetch API, and also remove URLs and have them be part of the .env file
 
 const axios = require('axios');
-require('dotenv').config();
+
 const { CERTIFICATION_API_KEY, ORGS_DATA_URL, SYSTEMS_DATA_URL } = process.env;
 const { sleep } = require('../../../common');
 
 const API_DEBOUNCE_SECONDS = 0.1;
 
-const postDataDictionaryResultsToApi = async ({
-  url,
-  providerUoi,
-  providerUsi,
-  recipientUoi,
-  metadataReport = {}
-} = {}) => {
+const postDataDictionaryResultsToApi = async ({ url, providerUoi, providerUsi, recipientUoi, metadataReport = {} } = {}) => {
   if (!url) throw new Error('url is required!');
   if (!providerUoi) throw new Error('providerUoi is required!');
   if (!providerUsi) throw new Error('providerUsi is required!');
@@ -25,20 +19,16 @@ const postDataDictionaryResultsToApi = async ({
   try {
     const { id: reportId = null } =
       (
-        await axios.post(
-          `${url}/api/v1/certification_reports/data_dictionary/${providerUoi}`,
-          metadataReport,
-          {
-            maxContentLength: Infinity,
-            maxBodyLength: Infinity,
-            headers: {
-              Authorization: `ApiKey ${CERTIFICATION_API_KEY}`,
-              recipientUoi,
-              'Content-Type': 'application/json',
-              providerUsi
-            }
+        await axios.post(`${url}/api/v1/certification_reports/data_dictionary/${providerUoi}`, metadataReport, {
+          maxContentLength: Infinity,
+          maxBodyLength: Infinity,
+          headers: {
+            Authorization: `ApiKey ${CERTIFICATION_API_KEY}`,
+            recipientUoi,
+            'Content-Type': 'application/json',
+            providerUsi
           }
-        )
+        })
       ).data || {};
 
     if (!reportId) throw new Error('Did not receive the required id parameter from the response!');
@@ -113,11 +103,7 @@ const getOrgsMap = async () => {
   if (!orgs?.length) throw new Error('ERROR: could not fetch Org data!');
 
   return orgs.reduce((acc, { OrganizationUniqueId, OrganizationType }) => {
-    if (
-      !['MLS', 'Technology Company', 'Pooled Platform', 'Commercial', 'Brokerage'].find(
-        type => type?.trim() === OrganizationType
-      )
-    ) {
+    if (!['MLS', 'Technology Company', 'Pooled Platform', 'Commercial', 'Brokerage'].find(type => type?.trim() === OrganizationType)) {
       return acc;
     }
